move spectramergin code
