+++ src/openms/include/OpenMS/FILTERING/SPECTRAMERGING/sources.cmake
@@ -1,14 +1,9 @@
 ### the directory name
-set(directory include/OpenMS/FILTERING/TRANSFORMERS)
+set(directory include/OpenMS/FILTERING/SPECTRAMERGING)
 
 ### list all header files of the directory here
 set(sources_list_h
-NLargest.h
-Normalizer.h
-RankScaler.h
-SqrtScaler.h
-ThresholdMower.h
-WindowMower.h
+SpectraMerger.h
 )
 
 ### add path to the filenames
@@ -18,7 +13,7 @@ foreach(i ${sources_list_h})
 endforeach(i)
 
 ### source group definition
-source_group("Header Files\\OpenMS\\FILTERING\\TRANSFORMERS" FILES ${sources_h})
+source_group("Header Files\\OpenMS\\FILTERING\\SPECTRAMERGING" FILES ${sources_h})
 
 set(OpenMS_sources_h ${OpenMS_sources_h} ${sources_h})
