+++ src/openms/source/FILTERING/SPECTRAMERGING/sources.cmake
@@ -1,14 +1,9 @@
 ### the directory name
-set(directory source/FILTERING/TRANSFORMERS)
+set(directory source/FILTERING/SPECTRAMERGING)
 
 ### list all filenames of the directory here
 set(sources_list
-NLargest.cpp
-Normalizer.cpp
-RankScaler.cpp
-SqrtScaler.cpp
-ThresholdMower.cpp
-WindowMower.cpp
+SpectraMerger.cpp
 )
 
 ### add path to the filenames
@@ -21,5 +16,5 @@ endforeach(i)
 set(OpenMS_sources ${OpenMS_sources} ${sources})
 
 ### source group definition
-source_group("Source Files\\FILTERING\\TRANSFORMERS" FILES ${sources})
+source_group("Source Files\\FILTERING\\SPECTRAMERGING" FILES ${sources})
