+++ src/openms/source/FORMAT/GNPSMGFFile.cpp
@@ -11,7 +11,7 @@
 #include <OpenMS/FORMAT/GNPSMGFFile.h>
 #include <OpenMS/COMPARISON/BinnedSpectralContrastAngle.h>
 #include <OpenMS/CONCEPT/UniqueIdInterface.h>
-#include <OpenMS/FILTERING/TRANSFORMERS/SpectraMerger.h>
+#include <OpenMS/FILTERING/SPECTRAMERGING/SpectraMerger.h>
 #include <OpenMS/FORMAT/FileHandler.h>
 #include <OpenMS/METADATA/PeptideIdentification.h>
 #include <OpenMS/KERNEL/OnDiscMSExperiment.h>
