+++ src/openms/source/IONMOBILITY/IMDataConverter.cpp
@@ -17,7 +17,7 @@
 
 
 #include <map>
-#include <OpenMS/FILTERING/TRANSFORMERS/SpectraMerger.h>
+#include <OpenMS/FILTERING/SPECTRAMERGING/SpectraMerger.h>
 
 namespace OpenMS
 {
